refactor: P2 server split 10/N — extract practice-lab route domain
Move /api/terminal/sessions* (create/close) and /api/runs* (list, detail,
archive-to-wiki) out of server.ts into server/routes/practice.ts as
handlePracticeRoutes. terminalManager/runRecordStore stay in server.ts
(the WebSocket PTY handler shares them) and are injected via ctx;
serializeFrontmatter's import leaves server.ts with the archive route.
server.ts: 2382 → 2267 lines.

Smoke coverage: runs list 400 without sessionId, run detail 404, terminal
create 400 without sessionId.
